// src/interfaces/EventTodayInterface.ts
export interface EventTodayInterface {
  id: number;
  title_bm: string;
  title_eng: string;
  description_bm: string;
  description_eng: string;
  script_date: string;
  day: string;
  script_day: string;
  script_month: string;
}

export interface AllEventsTodayInterface {
  no: number;
  day: string;
  title_bm: string;
  title_eng: string;
  script_date: string;
  description_bm: string;
  descriptiton_eng: string;
  id: number;
}